refactor: decouple repositories of mongodb done
