+++ internal/repository/mongo/user_repository.go
@@ -1,8 +1,10 @@
-package repository
+package repository_mongo
 
 import (
 	"context"
 	"errors"
+
+	"github.com/kavkaco/Kavka-Core/internal/repository"
 
 	"github.com/kavkaco/Kavka-Core/database"
 	"github.com/kavkaco/Kavka-Core/internal/model"
@@ -10,28 +12,11 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
-var (
-	ErrUserNotFound      = errors.New("user not found")
-	ErrEmailAlreadyTaken = errors.New("email already taken")
-)
-
-type UserRepository interface {
-	GetChats(ctx context.Context, userID model.UserID) ([]model.ChatID, error)
-	Create(ctx context.Context, user *model.User) (*model.User, error)
-	AddToUserChats(ctx context.Context, userID model.UserID, chatID model.ChatID) error
-	Update(ctx context.Context, userID string, name, lastName, username, biography string) error
-	FindOne(ctx context.Context, filter bson.M) (*model.User, error)
-	FindMany(ctx context.Context, filter bson.M) ([]*model.User, error)
-	FindByUserID(ctx context.Context, userID model.UserID) (*model.User, error)
-	FindByUsername(ctx context.Context, username string) (*model.User, error)
-	FindByEmail(ctx context.Context, email string) (*model.User, error)
-}
-
 type userRepository struct {
 	usersCollection *mongo.Collection
 }
 
-func NewUserRepository(db *mongo.Database) UserRepository {
+func NewUserMongoRepository(db *mongo.Database) repository.UserRepository {
 	return &userRepository{db.Collection(database.UsersCollection)}
 }
 
@@ -52,7 +37,7 @@ func (repo *userRepository) Update(ctx context.Context, userID string, name, las
 	}
 
 	if result.MatchedCount == 0 || result.ModifiedCount == 0 {
-		return ErrNotModified
+		return repository.ErrNotModified
 	}
 
 	return nil
@@ -84,7 +69,7 @@ func (repo *userRepository) GetChats(ctx context.Context, userID model.UserID) (
 func (repo *userRepository) Create(ctx context.Context, userModel *model.User) (*model.User, error) {
 	_, err := repo.usersCollection.InsertOne(context.TODO(), userModel)
 	if database.IsDuplicateKeyError(err) {
-		return nil, ErrEmailAlreadyTaken
+		return nil, repository.ErrEmailAlreadyTaken
 	} else if err != nil {
 		return nil, err
 	}
@@ -97,7 +82,7 @@ func (repo *userRepository) FindOne(ctx context.Context, filter bson.M) (*model.
 
 	result := repo.usersCollection.FindOne(context.TODO(), filter)
 	if errors.Is(result.Err(), mongo.ErrNoDocuments) {
-		return nil, ErrUserNotFound
+		return nil, repository.ErrUserNotFound
 	} else if result.Err() != nil {
 		return nil, result.Err()
 	}
@@ -108,24 +93,6 @@ func (repo *userRepository) FindOne(ctx context.Context, filter bson.M) (*model.
 	}
 
 	return model, nil
-}
-
-func (repo *userRepository) FindMany(ctx context.Context, filter bson.M) ([]*model.User, error) {
-	cursor, err := repo.usersCollection.Find(context.TODO(), filter)
-	if err != nil {
-		return nil, err
-	}
-
-	var users []*model.User
-
-	err = cursor.All(ctx, &users)
-	if errors.Is(err, mongo.ErrNoDocuments) {
-		return nil, ErrUserNotFound
-	} else if err != nil {
-		return nil, err
-	}
-
-	return users, nil
 }
 
 func (repo *userRepository) FindByUserID(ctx context.Context, userID model.UserID) (*model.User, error) {
