+++ tests/integration/service/chat_service_test.go
@@ -8,6 +8,7 @@ import (
 	lorem "github.com/bozaro/golorem"
 	"github.com/kavkaco/Kavka-Core/internal/model"
 	"github.com/kavkaco/Kavka-Core/internal/repository"
+	repository_mongo "github.com/kavkaco/Kavka-Core/internal/repository/mongo"
 	service "github.com/kavkaco/Kavka-Core/internal/service/chat"
 	"github.com/kavkaco/Kavka-Core/utils"
 	"github.com/kavkaco/Kavka-Core/utils/random"
@@ -31,8 +32,8 @@ type ChatTestSuite struct {
 func (s *ChatTestSuite) SetupSuite() {
 	s.lem = lorem.New()
 
-	chatRepo := repository.NewChatRepository(db)
-	userRepo := repository.NewUserRepository(db)
+	chatRepo := repository_mongo.NewChatMongoRepository(db)
+	userRepo := repository_mongo.NewUserMongoRepository(db)
 
 	s.userRepo = userRepo
 	s.service = service.NewChatService(chatRepo, userRepo)
