+++ tests/integration/service/user_service_test.go
@@ -1,1 +1,4 @@
 package service
+
+// TODO
+// Write integration tests for user!
